filter by athlete

+++ data-storage/src/main/scala/velocorner/model/Account.scala
@@ -26,7 +26,7 @@ object Account {
 }
 
 case class Account(
-  athleteId: Long,
+  athleteId: Int,
   displayName: String,
   displayLocation: String,
   avatarUrl: String,
